Adjust upload test with file format check
RISDEV-175
